// src/store/main/userOperation/types.ts
/*
 * @Author: korealu
 * @Date: 2022-02-16 16:53:07
 * @LastEditors: Please set LastEditors
 * @LastEditTime: 2022-03-17 09:49:27
 * @Description: file content
 * @FilePath: /pofi-admin/src/store/main/finance/pay/types.ts
 */
export interface IUserOperationType {
  userOperationsList: any[]
  userOperationsCount: number
  logList: any[]
  logCount: number
}